pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.16",
    "celery>=5.5.3",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "gevent>=25.5.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.3.0",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
